fix(web): move agent work tabs into page header
Keep the Automations and Tasks switcher in the shared navigation header so the unified page avoids a redundant content toolbar.
